EXAMSYS-3531 Calculate the size of equations accurately
The way the code tried to calculate the height and width before this change was completely wrong.

After this change we use the browser to calculate now much space an element takes up using scrollWidth and scrollHeight. Before we get these values we set the element height and width to 0, because if we do not and the element already has a value set that is greater than the space needed it will return that value.

The updateMEE function is no longer used after these changes.